Jan branch 4

+++ Tophat-Railway/src/application/TrainModel/TrainModel.java
@@ -291,7 +291,7 @@ class TrainModel extends JPhysics implements TrainInterface, TrainModelTrackInte
 		return acceleration;
 	}
 
-	double getCrew() {
+	int getCrew() {
 		return crewCount;
 	}
